w3 trunk
  Undo of r21153 commit. The version of w3 on which the changes were
  made by GeorgeV were an earlier version - May2011 - that did not
  contain a number of NCO/SIB changes made in Nov2011 and Jan2012.
  See nceplibs Trac Ticket #26.
  https://svnemc.ncep.noaa.gov/trac/nceplibs/ticket/26

+++ libsrc/w3ai15.f
@@ -63,7 +63,7 @@ C
       INTEGER     IDIV(8)
       INTEGER     NBUFA(*)
       INTEGER     NBUFB(*)
-      INTEGER*8   ZERO(8)
+      INTEGER     ZERO(8)
 C
       CHARACTER*1 BLANK
       CHARACTER*1 JTEMP(8)
